+++ src/store/telegramStore.types.ts
@@ -139,6 +139,7 @@ export interface TelegramState {
   ) => Promise<ForwardMessagesResult | undefined>;
   cacheFile: (fileId: number, priority?: number) => Promise<void>;
   streamFile: (fileId: number, size: number, mimeType?: string) => Promise<string | undefined>;
+  suspendFileStream: (fileId: number) => Promise<void>;
   downloadFile: (fileId: number, fileName: string) => Promise<void>;
   cancelFileDownload: (fileId: number) => Promise<void>;
   openFile: (sourcePath: string) => Promise<void>;